% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_suite.R
\name{benchmark_suite_retrieve}
\alias{benchmark_suite_retrieve}
\title{Downloading and processing the results of the benchmark jobs}
\usage{
benchmark_suite_retrieve(out_dir, return_outputs = TRUE)
}
\arguments{
\item{out_dir}{The folder in which to output intermediate and final results.}

\item{return_outputs}{Whether or not to return the outputs, or only retrieve them from the cluster.}
}
\description{
Downloading and processing the results of the benchmark jobs
}
