% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_bind_results.R
\name{bs_bind_results}
\alias{bs_bind_results}
\title{Gather and bind the results of the benchmark jobs}
\usage{
bs_bind_results(out_dir, load_models = FALSE)
}
\arguments{
\item{out_dir}{The folder in which to output intermediate and final results.}

\item{load_models}{Whether or not to load the models as well.}
}
\description{
Gather and bind the results of the benchmark jobs
}
