% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_fetch_results.R
\name{bs_fetch_results}
\alias{bs_fetch_results}
\title{Fetch the results of the benchmark jobs from the cluster.}
\usage{
bs_fetch_results(out_dir)
}
\arguments{
\item{out_dir}{The folder in which to output intermediate and final results.}
}
\description{
Fetch the results of the benchmark jobs from the cluster.
}
