% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{cal_branch_assignment}
\alias{cal_branch_assignment}
\title{Convert percentages to a unique edge assignment}
\usage{
cal_branch_assignment(percentages, network)
}
\description{
Convert percentages to a unique edge assignment
}
