% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_branch_overlap.R
\name{calculate_branch_overlap}
\alias{calculate_branch_overlap}
\title{Compares the overlap between branches}
\usage{
calculate_branch_overlap(dataset, prediction)
}
\arguments{
\item{dataset}{A dataset}

\item{prediction}{A predicted model}
}
\description{
Compares the overlap between branches
}
