% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_combinations.R
\name{calculate_harmonic_mean}
\alias{calculate_harmonic_mean}
\alias{calculate_geometric_mean}
\alias{calculate_arithmetic_mean}
\title{Calculate combinations of metrics}
\usage{
calculate_harmonic_mean(..., weights = NULL)

calculate_geometric_mean(..., weights = NULL)

calculate_arithmetic_mean(..., weights = NULL)
}
\arguments{
\item{...}{Can be:
\itemize{
\item One numeric vector
\item A list containg numeric vectors
\item Numeric vectors given as separate inputs
}}

\item{weights}{Weights with the same length as ...}
}
\description{
Calculate combinations of metrics
}
\examples{
calculate_harmonic_mean(0.1, 0.5, 0.9)
calculate_geometric_mean(0.1, 0.5, 0.9)
calculate_arithmetic_mean(0.1, 0.5, 0.9)

calculate_harmonic_mean(c(0.1, 0.9), c(0.2, 1))
calculate_arithmetic_mean(c(0.1, 10), c(0.9, 20))
}
