% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_flip.R
\name{calculate_edge_flip}
\alias{calculate_edge_flip}
\title{Edge flip score}
\usage{
calculate_edge_flip(net1, net2, return = c("score", "all"), simplify = TRUE,
  limit_flips = 5, limit_combinations = choose(25, 4))
}
\arguments{
\item{net1}{Network 1}

\item{net2}{Network 2}

\item{return}{Whether to return only the `score` or the full output (`all`)}

\item{simplify}{Whether or not to simplify the networks}

\item{limit_flips}{Maximal number of flips to check}

\item{limit_combinations}{Maximal number of combinations to check}
}
\description{
Edge flip score
}
\examples{
net1 <- dyntoy::generate_milestone_network("linear")
net2 <- dyntoy::generate_milestone_network("bifurcating")
calculate_edge_flip(net1, net2)

net1 <- dyntoy::generate_milestone_network("cyclic")
net2 <- dyntoy::generate_milestone_network("diverging_with_loops")
calculate_edge_flip(net1, net2)

}
