% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_featureimp.R
\name{calculate_featureimp_cor}
\alias{calculate_featureimp_cor}
\title{Compare feature importances derived by both trajectories}
\usage{
calculate_featureimp_cor(dataset, prediction, num_trees = 10000,
  mtry = function(x) ncol(x) * 0.01)
}
\arguments{
\item{dataset}{A dataset}

\item{prediction}{A predicted model}

\item{num_trees}{The number of trees to use for the random forest}

\item{mtry}{Number of features to split in each node. Can be a function with as argument the dataset}
}
\description{
Compare feature importances derived by both trajectories
}
