% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{calculate_ged}
\alias{calculate_ged}
\title{Compute the graph edit distance using gedevo}
\usage{
calculate_ged(net1, net2)
}
\arguments{
\item{net1}{the first network to compare}

\item{net2}{the second network to compare}
}
\description{
Compute the graph edit distance using gedevo
}
