% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_him.R
\name{calculate_him}
\alias{calculate_him}
\title{netdist scores}
\usage{
calculate_him(net1, net2, simplify = TRUE)
}
\arguments{
\item{net1}{Network 1}

\item{net2}{Network 2}

\item{simplify}{Whether or not to simplify the networks}
}
\description{
netdist scores
}
\examples{
net1 <- dyntoy::generate_milestone_network("linear")
net2 <- dyntoy::generate_milestone_network("bifurcating")
calculate_him(net1, net2)

net1 <- dyntoy::generate_milestone_network("cyclic")
net2 <- dyntoy::generate_milestone_network("cyclic")
calculate_him(net1, net2)

}
\keyword{metric}
