% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_mapping.R
\name{calculate_mapping}
\alias{calculate_mapping}
\alias{calculate_mapping_milestones}
\alias{calculate_mapping_branches}
\title{Compares the mapping between milestones or branches}
\usage{
calculate_mapping(dataset, prediction, grouping = c("branches",
  "milestones"), simplify = TRUE)

calculate_mapping_milestones(dataset, prediction, simplify = TRUE)

calculate_mapping_branches(dataset, prediction, simplify = TRUE)
}
\arguments{
\item{dataset}{A dataset}

\item{prediction}{A predicted model}

\item{grouping}{How to group the cells, either branches or milestones}

\item{simplify}{Whether to simplify the trajectory (allowing self loops)}
}
\description{
Compares the mapping between milestones or branches
}
\keyword{metric}
