% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate the similarity between two trajectory models.}
\usage{
calculate_metrics(dataset, model, metrics = dyneval::metrics$metric_id,
  expression_source = dataset$expression)
}
\arguments{
\item{dataset}{The first trajectory, in most cases a gold standard trajectory}

\item{model}{The second trajectory, in most cases a predicted trajectory}

\item{metrics}{Which metrics to evaluate. Check \code{dyneval::metrics} for a list of possible metrics.
Passing a custom metric function with format \code{function(dataset, model) { 1 }} is also supported. The name of this function within the list will be used as the name of the metric.}

\item{expression_source}{The expression data matrix, with features as columns.
\itemize{
\item If a matrix is provided, it is used as is.
\item If a character is provided, \code{dataset[[expression_source]]} should contain the matrix.
\item If a function is provided, that function will be called in order to obtain the expression (useful for lazy loading).
}}
}
\description{
One use case for these metrics is to calculate the accuracy of a certain prediction compared to a reference trajectory. However, these metrics can also be used for other purposes, such as clustering of trajectories.
}
\details{
Some metrics are asymmetric (see \code{dyneval::metrics$symmetric}), in which case the order of the dataset and model parameters matters.
}
\keyword{metric}
