% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_node_edge_edit_score.R
\name{calculate_node_edge_edit_score}
\alias{calculate_node_edge_edit_score}
\title{Compute node edge edit score}
\usage{
calculate_node_edge_edit_score(net1, net2)
}
\arguments{
\item{net1}{the first network to compare}

\item{net2}{the second network to compare}
}
\description{
Compute node edge edit score
}
