% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_node_edit_score.R
\name{calculate_node_edit_score}
\alias{calculate_node_edit_score}
\title{Compute the node edit score}
\usage{
calculate_node_edit_score(net1, net2, nodes1 = unique(c(net1$from, net1$to)),
  nodes2 = unique(c(net2$from, net2$to)), normalize_weights = TRUE)
}
\arguments{
\item{net1}{the first network to compare}

\item{net2}{the second network to compare}

\item{nodes1}{nodes in `net1`, defaults to unique from and to}

\item{nodes2}{nodes in `net2`, defaults to unique from and to}

\item{normalize_weights}{Whether to normalize the lengths of each network}
}
\description{
Compute the node edit score
}
