% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_position_predict.R
\name{calculate_position_predict}
\alias{calculate_position_predict}
\title{Compute metrics related to the prediction of the positions}
\usage{
calculate_position_predict(dataset, prediction, metrics = c("rf_mse",
  "rf_rsq", "lm_mse", "lm_rsq"))
}
\arguments{
\item{dataset}{A dataset containing a trajectory}

\item{prediction}{A predicted trajectory}

\item{metrics}{Which metrics to predict, can be rf_mse, rf_rsq, rf_nmse, lm_mse, lm_rsq and/or lm_nmse}
}
\description{
Compute metrics related to the prediction of the positions
}
\keyword{metric}
