% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_coranking.R
\name{compute_coranking}
\alias{compute_coranking}
\title{Compute the coranking matrix and}
\usage{
compute_coranking(gold_dist, pred_dist)
}
\arguments{
\item{gold_dist}{A data frame containing the pairwise distances in the original space}

\item{pred_dist}{A data frame containing the pairwise distances in the new space}
}
\description{
Compute the coranking matrix and
}
