% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_featureimp.R
\name{compute_fimp_ks}
\alias{compute_fimp_ks}
\title{Compare enrichment in finding back the most important genes}
\usage{
compute_fimp_ks(dataset, prediction, num_trees = 10000)
}
\arguments{
\item{dataset}{A dataset}

\item{prediction}{A predicted model}

\item{num_trees}{the number of trees to use during the calculation of the metric}
}
\description{
Compare enrichment in finding back the most important genes
}
