% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_rfmse.R
\name{compute_rfmse}
\alias{compute_rfmse}
\title{Compute the random forest OOB-MSE metric}
\usage{
compute_rfmse(task, prediction)
}
\arguments{
\item{task}{A task}

\item{prediction}{A predicted model}
}
\description{
Compute the random forest OOB-MSE metric
}
