% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_method.R
\name{create_description}
\alias{create_description}
\title{Create a new TI description}
\usage{
create_description(name, short_name, package_loaded, package_required, par_set,
  properties, run_fun, plot_fun, make_command = NULL)
}
\description{
Create a new TI description
}
