% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scuba.R, R/ti_scuba_py.R
\name{description_scuba}
\alias{description_scuba}
\alias{description_scuba}
\title{Description for SCUBA}
\usage{
description_scuba()

description_scuba()
}
\description{
Description for SCUBA

Description for SCUBA
}
