% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{dimensionality_reduce_trajectory}
\alias{dimensionality_reduce_trajectory}
\title{Perform dimensionality reduction on a trajectory and the
respective cells}
\usage{
dimensionality_reduce_trajectory(name, traj)
}
\arguments{
\item{name}{The name of the trajectory (mandatory)}

\item{traj}{The trajectory itself. TODO: explain format}
}
\description{
Perform dimensionality reduction on a trajectory and the
respective cells
}
