% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{emdist}
\alias{emdist}
\title{Calculate Earth Mover's Distance between cells in a trajectory}
\usage{
emdist(traj)
}
\arguments{
\item{traj}{the trajectory}
}
\description{
Calculate Earth Mover's Distance between cells in a trajectory
}
