% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{execute_evaluation}
\alias{execute_evaluation}
\title{Running an evaluation of a method on a set of tasks with a set of parameters}
\usage{
execute_evaluation(tasks, method, parameters, metrics, timeout,
  output_model = TRUE, error_score = -1)
}
\arguments{
\item{tasks}{The tasks on which to evaluate.}

\item{method}{The method to evaluate.}

\item{parameters}{The parameters to evaluate with.}

\item{metrics}{which metrics to use;
see \code{\link{calculate_metrics}} for a list of which metrics are available.}

\item{timeout}{Kill execution after a given amount of time.}

\item{output_model}{Whether or not the model will be outputted.
If this is a character string, it will save the model in the requested folder.}

\item{error_score}{The aggregated score a method gets if it produces errors.}
}
\description{
Running an evaluation of a method on a set of tasks with a set of parameters
}
