% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_method.R
\name{execute_method_internal}
\alias{execute_method_internal}
\title{Internal method for executing a method}
\usage{
execute_method_internal(tasks, method, parameters, give_start_cell,
  give_end_cells, give_cell_grouping)
}
\arguments{
\item{tasks}{The tasks on which to evaluate.}

\item{method}{The method to evaluate.}

\item{parameters}{The parameters to evaluate with.}

\item{give_start_cell}{Whether a start cell should be provided even though a method doesn't require it.}

\item{give_end_cells}{Whether end cells should be provided even though a method doesn't require it.}

\item{give_cell_grouping}{Whether a cell grouping should be provided even though a method doesn't require it.}
}
\description{
If you're reading this, you're supposed to be using \code{execute_method} instead.
}
