% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_row_to_list}
\alias{extract_row_to_list}
\title{Extracts one row, and reduces list-columns to their contents}
\usage{
extract_row_to_list(df, row_id)
}
\description{
Extracts one row, and reduces list-columns to their contents
}
