% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_parameters.R
\name{get_parameter_row}
\alias{get_parameter_row}
\title{Return a set of parameter values from the expanded parameter grid}
\usage{
get_parameter_row(method, parameter_row)
}
\arguments{
\item{method}{the wrapped method to extract the expanded parameters from}

\item{parameter_row}{the index of the expanded parameters to use}
}
\description{
Return a set of parameter values from the expanded parameter grid
}

