% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{impute_y_fun}
\alias{impute_y_fun}
\title{For returning a poor score when a method errors}
\usage{
impute_y_fun(num_objectives, error_score = -1)
}
\arguments{
\item{num_objectives}{the number of objectives used}

\item{error_score}{the score to return upon erroring}
}
\description{
For returning a poor score when a method errors
}
