% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_types.R
\name{is_wrapped_data_object}
\alias{is_wrapped_data_object}
\title{Returns whether an object is a data_object}
\usage{
is_wrapped_data_object(object)
}
\arguments{
\item{object}{the object to test}
}
\description{
Returns whether an object is a data_object
}

