% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{make_obj_fun}
\alias{make_obj_fun}
\title{Used for wrapping an evaluation function around a TI method}
\usage{
make_obj_fun(method, noisy = F, load_packages = T, suppress_output = T,
  metrics = c("Q_global", "Q_local", "correlation"))
}
\description{
Used for wrapping an evaluation function around a TI method
}
