% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{make_trajectory_plot}
\alias{make_trajectory_plot}
\title{Plot a dimensionality reduced trajectory}
\usage{
make_trajectory_plot(dimred_traj)
}
\arguments{
\item{dimred_traj}{output as generated by \code{dimensionality_reduce_trajectory}}
}
\description{
Plot a dimensionality reduced trajectory
}
