% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{Metadata on metrics}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 10 columns.}
\usage{
metrics
}
\description{
Also contained are:
\itemize{
\item Ways of rendering the metrics
\item Categorisation of the metrics
\item The perfect and worst possible scores (if possible to determine beforehand)
\item Whether the metric is symmetric (i.e. whether flipping the order of trajectories would change the expected value of the metric)
}
}
\keyword{datasets}
