% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{my_assignin_namespace}
\alias{my_assignin_namespace}
\title{This function assigns an object in a different namespace.}
\usage{
my_assignin_namespace(x, value, ns, pos = -1, envir = as.environment(pos))
}
\description{
It was copied from \code{\link[utils]{assignInNamespace}},
but without an extra check so that I can override the set.seed function.
}
