% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods_trajectory.R
\name{my_read_ods}
\alias{my_read_ods}
\title{read data from ods files}
\usage{
my_read_ods(path = NULL, sheet = 1, col_names = TRUE, col_types = NULL,
  na = "", skip = 0, formula_as_formula = FALSE, range = NULL)
}
\description{
Bugfix wrapper of the \code{\link[readODS]{read_ods}} function
}
