% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_default}
\alias{plot_default}
\title{Plot a dimensionality reduced trajectory}
\usage{
plot_default(object, insert_phantom_edges = T)
}
\arguments{
\item{object}{output as generated by \code{plotLearnerData.ti.default}, \link{\code{wrap_ti_task_data}} or \link{\code{wrap_ti_prediction}}}

\item{insert_phantom_edges}{attempt to plot bifurcating edges correctly automatically}
}
\description{
Plot a dimensionality reduced trajectory
}
