% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlearner_ti_plot.R
\name{plotLearnerData.ti.default}
\alias{plotLearnerData.ti.default}
\title{Perform dimensionality reduction on a trajectory and the respective samples in order to plot it}
\usage{
plotLearnerData.ti.default(traj_object, insert_phantom_edges = T)
}
\arguments{
\item{traj_object}{A trajectory as generated by \link{\code{wrap_ti_task_data}} or
\link{\code{wrap_ti_prediction}}}

\item{insert_phantom_edges}{attempt to plot bifurcating edges correctly automatically}
}
\description{
Perform dimensionality reduction on a trajectory and the respective samples in order to plot it
}
