% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{process_dyneval_plot}
\alias{process_dyneval_plot}
\title{Default theme for TI plots}
\usage{
process_dyneval_plot(g, id, expand = TRUE)
}
\arguments{
\item{g}{A ggplot to modify}

\item{id}{The title}

\item{expand}{Whether or not to leave space at the borders}
}
\description{
Default theme for TI plots
}
