% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods_trajectory.R
\name{read_ods_trajectory}
\alias{read_ods_trajectory}
\title{Read an ODS file containing a trajectory. TODO: explain format. Ask Robrecht about the format for now.}
\usage{
read_ods_trajectory(filename)
}
\arguments{
\item{filename}{the location of the ODS}
}
\description{
Read an ODS file containing a trajectory. TODO: explain format. Ask Robrecht about the format for now.
}
