% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlearner_ti_wrappers.R
\name{makeRLearnerTI}
\alias{makeRLearnerTI}
\title{TI R Learner}
\usage{
makeRLearnerTI(cl, package, par.set, par.vals = list(),
  properties = character(0L), name = cl, short.name = cl, note = "",
  callees = character(0L))
}
\description{
TI R Learner
}
