% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_method_types.R
\name{run_method}
\alias{run_method}
\title{Running a dyneval method}
\usage{
run_method(wrapped_method, data, ...)
}
\arguments{
\item{wrapped_method}{The wrapped method}

\item{data}{The wrapped data to be passed}

\item{...}{Parameters of the method}
}
\description{
Running a dyneval method
}

