% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_wishbone.R
\name{run_wishbone}
\alias{run_wishbone}
\title{Run Wishbone}
\usage{
run_wishbone(counts, start_cell_id, knn = 10, n_diffusion_components = 2,
  n_pca_components = 15, markers = "~", branch = TRUE, k = 15,
  num_waypoints = 50, normalize = TRUE, epsilon = 1)
}
\description{
Run Wishbone
}
