% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_flip.R
\name{score_edge_flips}
\alias{score_edge_flips}
\title{Edge flip score}
\usage{
score_edge_flips(net1, net2, return = c("score", "all"), simplify = TRUE)
}
\arguments{
\item{net1}{Network 1}

\item{net2}{Network 2}

\item{return}{Whether to return only the `score` or the full output (`all`)}
}
\description{
Edge flip score
}
