% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{score_prediction_F1rr}
\alias{score_prediction_F1rr}
\title{Compute the F1rr score for overlap of branches}
\usage{
score_prediction_F1rr(task, pred_output)
}
\description{
Can currently not handle multiple linear edges without branching, which should be merged in the future
}
