% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_types.R
\name{unwrap_data_object}
\alias{unwrap_data_object}
\title{Unwrap a data_object}
\usage{
unwrap_data_object(wrapped_data_object)
}
\arguments{
\item{wrapped_data_object}{a wrapped data_object}
}
\description{
Unwrap a data_object
}

