% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_types.R
\name{wrap_data_object}
\alias{wrap_data_object}
\title{Wrap a data object}
\usage{
wrap_data_object(data_type_object, data_object)
}
\arguments{
\item{data_object}{the data itself}

\item{data_types}{the data type object}
}
\description{
Wrap a data object
}

