% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_data.R
\name{wrap_ti_prediction}
\alias{wrap_ti_prediction}
\title{Wrapping TI predictions}
\usage{
wrap_ti_prediction(ti_type, id, cell_ids, milestone_ids, milestone_network,
  milestone_percentages = NULL, progressions = NULL, ...)
}
\description{
Wrapping TI predictions
}
