% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_data.R
\name{wrap_ti_task_data}
\alias{wrap_ti_task_data}
\title{Wrapping TI task data}
\usage{
wrap_ti_task_data(ti_type, id, counts, cell_ids, milestone_ids,
  milestone_network, milestone_percentages = NULL, progressions = NULL,
  sample_info = NULL, feature_info = NULL, ...)
}
\description{
Wrapping TI task data
}
