% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_suite.R
\name{benchmark_suite_submit}
\alias{benchmark_suite_submit}
\title{A benchmark suite with which to run all the methods on the different tasks}
\usage{
benchmark_suite_submit(tasks, task_group, task_fold, out_dir,
  methods = get_descriptions(as_tibble = F), num_cores = 4,
  metrics = c("auc_R_nx", "robbie_network_score"), num_iterations = 20,
  num_init_params = 100, num_repeats = 1)
}
\arguments{
\item{tasks}{A tibble of tasks}

\item{task_group}{A grouping vector for the different tasks}

\item{task_fold}{A fold index vector for the different tasks}

\item{out_dir}{The folder in which to output intermediate and final results}

\item{num_cores}{How many cores to use per mlrMBO process}

\item{metrics}{Which metrics to use}

\item{num_iterations}{The number of iterations to run}

\item{num_init_params}{The number of initial parameters to evaluate}

\item{num_repeats}{The number of times to repeat the mlr process, for each group and each fold}
}
\description{
A benchmark suite with which to run all the methods on the different tasks
}
