% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_flip.R
\name{calculate_edge_flip}
\alias{calculate_edge_flip}
\title{Edge flip score}
\usage{
calculate_edge_flip(net1, net2, return = c("score", "all"), simplify = TRUE)
}
\arguments{
\item{net1}{Network 1}

\item{net2}{Network 2}

\item{return}{Whether to return only the `score` or the full output (`all`)}

\item{simplify}{Whether or not to simplify the networks}
}
\description{
Edge flip score
}
