% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_featureimp.R
\name{calculate_featureimp_cor}
\alias{calculate_featureimp_cor}
\title{Compare feature importances derived by both trajectories}
\usage{
calculate_featureimp_cor(dataset, prediction,
  fi_method = dynfeature::fi_ranger_rf_lite())
}
\arguments{
\item{dataset}{A dataset}

\item{prediction}{A predicted model}

\item{fi_method}{A feature importance method. Default: `fi_ranger_rf_lite()`. Check `?fi_methods` for a full list of available feature importance methods.}
}
\description{
Compare feature importances derived by both trajectories
}
