% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate the performance of a model with respect to a task}
\usage{
calculate_metrics(task, model, metrics)
}
\arguments{
\item{task}{the original task}

\item{model}{the predicted model}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Coranking of geodesic distances: \code{"mean_R_nx"}, \code{"auc_R_nx"}, \code{"Q_local"}, \code{"Q_global"}
  \item Spearman correlation of geodesic distances: \code{"correlation"}
  \item Mantel test p-value: \code{"mantel_pval"}
  \item Edge flip score: \code{"edge_flip"}
  \item RF MSE: \code{"rf_mse"}, \code{"rf_rsq"}
}}
}
\description{
Calculate the performance of a model with respect to a task
}
