% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate the performance of a model with respect to a dataset}
\usage{
calculate_metrics(dataset, model, metrics = dyneval::metrics$metric_id,
  expression_source = dataset$expression)
}
\arguments{
\item{dataset}{The original dataset.}

\item{model}{The predicted model.}

\item{metrics}{Which metrics to evaluate. Check \code{dyneval::metrics} for a list of possible metrics.
Passing a custom metric function with format \code{function(dataset, model) { 1 }} is also supported.}

\item{expression_source}{The expression data matrix, with features as columns.
\itemize{
\item If a matrix is provided, it is used as is.
\item If a character is provided, \code{dataset[[expression_source]]} should contain the matrix.
\item If a function is provided, that function will be called in order to obtain the expression (useful for lazy loading).
}}
}
\description{
Calculate the performance of a model with respect to a dataset
}
