% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate the performance of a model with respect to a dataset}
\usage{
calculate_metrics(dataset, model, metrics = c("correlation", "edge_flip",
  "isomorphic", "him", "rf_mse", "rf_rsq", "lm_mse", "lm_rsq",
  "featureimp_cor", "F1_branches", "F1_milestones"))
}
\arguments{
\item{dataset}{the original dataset}

\item{model}{the predicted model}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Spearman correlation of geodesic distances: \code{"correlation"}
  \item Edge flip score: \code{"edge_flip"}
  \item Isomorphic: \code{"isomorphic"}
  \item RF MSE: \code{"rf_mse"}, \code{"rf_rsq"}
  \item Similarity in feature importance: \code{"featureimp_cor"}
  \item Overlap between branches: \code{"F1_branches"}
  \item Custom metric function. Format: \code{function(dataset, model) { 1 }}
}}
}
\description{
Calculate the performance of a model with respect to a dataset
}
