% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate the performance of a model with respect to a task}
\usage{
calculate_metrics(task, model, metrics)
}
\arguments{
\item{task}{the original task}

\item{model}{the predicted model}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Coranking of geodesic distances: \code{"mean_R_nx"}, \code{"auc_R_nx"}, \code{"Q_local"}, \code{"Q_global"}
  \item Spearman correlation of geodesic distances: \code{"correlation"}
  \item Isomorphism of the two networks: \code{"isomorphic"}
  \item GEDEVO Graph Edit Distance: \code{"ged"}
  \item Earth Mover's Distance on orbit counts: \code{"net_emd"}
  \item Genetic Algorithm for aligning small graphs: \code{"robbie_network_score"}
}}
}
\description{
Calculate the performance of a model with respect to a task
}
