% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate the performance of a model with respect to a task}
\usage{
calculate_metrics(task, model, metrics = c("correlation", "edge_flip",
  "rf_mse", "rf_rsq", "featureimpcor"))
}
\arguments{
\item{task}{the original task}

\item{model}{the predicted model}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Spearman correlation of geodesic distances: \code{"correlation"}
  \item Edge flip score: \code{"edge_flip"}
  \item RF MSE: \code{"rf_mse"}, \code{"rf_rsq"}
  \item Correlation of feature importance scores
}}
}
\description{
Calculate the performance of a model with respect to a task
}
