% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{calculate_robbie_network_score}
\alias{calculate_robbie_network_score}
\title{Compute the robbie network score (any resemblances to real-life persons is purely coincidental)}
\usage{
calculate_robbie_network_score(net1, net2)
}
\arguments{
\item{net1}{the first network to compare}

\item{net2}{the second network to compare}
}
\description{
Compute the robbie network score (any resemblances to real-life persons is purely coincidental)
}
