% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_featureimp.R
\name{compute_featureimp}
\alias{compute_featureimp}
\title{Compare feature importances derived by both trajectories}
\usage{
compute_featureimp(dataset, prediction, num_trees = 50000)
}
\arguments{
\item{dataset}{A dataset}

\item{prediction}{A predicted model}

\item{num_trees}{the number of trees to use during the calculation of the metric}
}
\description{
Compare feature importances derived by both trajectories
}
