% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{execute_evaluation}
\alias{execute_evaluation}
\title{Running an evaluation of a method on a set of tasks with a set of parameters}
\usage{
execute_evaluation(tasks, method, parameters, metrics = c("mean_R_nx",
  "auc_R_nx", "Q_global", "Q_local", "correlation", "isomorphic",
  "robbie_network_score"), suppress_output = TRUE)
}
\arguments{
\item{tasks}{the tasks on which to evaluate}

\item{method}{the method to evaluate}

\item{parameters}{the parameters to evaluate with}

\item{metrics}{which metrics to use}

\item{suppress_output}{whether or not to suppress the outputted messages}
}
\description{
Running an evaluation of a method on a set of tasks with a set of parameters
}
