% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{execute_evaluation}
\alias{execute_evaluation}
\title{Running an evaluation of a method on a set of tasks with a set of parameters}
\usage{
execute_evaluation(tasks, method, parameters, metrics = c("mean_R_nx",
  "auc_R_nx", "Q_global", "Q_local", "correlation", "isomorphic",
  "robbie_network_score"))
}
\arguments{
\item{tasks}{The tasks on which to evaluate.}

\item{method}{The method to evaluate.}

\item{parameters}{The parameters to evaluate with.}

\item{metrics}{which metrics to use;
see \code{\link{calculate_metrics}} for a list of which metrics are available.}
}
\description{
Running an evaluation of a method on a set of tasks with a set of parameters
}
