% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_evaluation.R
\name{execute_evaluation}
\alias{execute_evaluation}
\title{Running an evaluation of a method on a set of tasks with a set of parameters}
\usage{
execute_evaluation(tasks, method, parameters, metrics, extra_metrics = NULL,
  output_model = TRUE, error_score = 0, mc_cores = 1)
}
\arguments{
\item{tasks}{The tasks on which to evaluate.}

\item{method}{The method to evaluate.}

\item{parameters}{The parameters to evaluate with.}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Coranking of geodesic distances: \code{"mean_R_nx"}, \code{"auc_R_nx"}, \code{"Q_local"}, \code{"Q_global"}
  \item Spearman correlation of geodesic distances: \code{"correlation"}
  \item Isomorphism of the two networks: \code{"isomorphic"}
  \item GEDEVO Graph Edit Distance: \code{"ged"}
  \item Earth Mover's Distance on orbit counts: \code{"net_emd"}
  \item Genetic Algorithm for aligning small graphs: \code{"robbie_network_score"}
  \item Mantel test p-value: \code{"mantel_pval"}
}}

\item{extra_metrics}{Extra metrics to calculate but not evaluate with.}

\item{output_model}{Whether or not the model will be outputted.
If this is a character string, it will save the model in the requested folder.}

\item{error_score}{The aggregated score a method gets if it produces errors.}

\item{mc_cores}{The number of cores to use, allowing to parallellise the different tasks}
}
\description{
Running an evaluation of a method on a set of tasks with a set of parameters
}
