% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_method.R
\name{execute_method}
\alias{execute_method}
\title{Run a method on a set of tasks with a set of parameters}
\usage{
execute_method(tasks, method, parameters, give_start_cell = FALSE,
  give_cell_grouping = FALSE, suppress_output = TRUE)
}
\arguments{
\item{tasks}{the tasks on which to evaluate}

\item{method}{the method to evaluate}

\item{parameters}{the parameters to evaluate with}

\item{give_start_cell}{whether a start cell should be provided even though a method doesn't require it}

\item{give_cell_grouping}{whether a cell grouping should be provided even though a method doesn't require it}

\item{suppress_output}{whether or not to suppress the outputted messages}
}
\description{
Run a method on a set of tasks with a set of parameters
}
