% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_method.R
\name{execute_method}
\alias{execute_method}
\title{Run a method on a set of tasks with a set of parameters}
\usage{
execute_method(tasks, method, parameters, give_start_cell = FALSE,
  give_end_cells = FALSE, give_grouping_assignment = FALSE, timeout = Inf,
  debug_timeout = FALSE)
}
\arguments{
\item{tasks}{The tasks on which to evaluate.}

\item{method}{The method to evaluate.}

\item{parameters}{The parameters to evaluate with.}

\item{give_start_cell}{Whether a start cell should be provided even though a method doesn't require it.}

\item{give_end_cells}{Whether end cells should be provided even though a method doesn't require it.}

\item{give_grouping_assignment}{Whether a cell grouping should be provided even though a method doesn't require it.}

\item{timeout}{Kill execution after a given amount of time.}

\item{debug_timeout}{Setting debug to \code{TRUE} will avoid running the method in a separate R session
using \code{\link[dynutils]{eval_with_timeout}} and run the method directly. Note that the timeout functionality
will not work when \code{debug} is \code{TRUE}.}
}
\description{
Run a method on a set of tasks with a set of parameters
}
