% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load_datasets}
\alias{load_datasets}
\title{Load dyngen datasets}
\usage{
load_datasets(mc_cores = 1, datasets_info = load_datasets_info())
}
\arguments{
\item{mc_cores}{The number of cores to use while loading in the datasets (default: 1)}

\item{datasets_info}{The meta information of the datasets to read}
}
\value{
A tibble of datasets
}
\description{
\strong{Important!} A variable \code{.datasets_location} needs to be defined in the global environment for this function
to work correctly.
}
\examples{
datasets <- load_datasets(mc_cores = 1, datasets_info = load_datasets_info()))
}
