% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{make_obj_fun}
\alias{make_obj_fun}
\title{Used for wrapping an evaluation function around a TI method}
\usage{
make_obj_fun(method, metrics, noisy = FALSE)
}
\arguments{
\item{method}{The method to evaluate.}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Coranking of geodesic distances: \code{"mean_R_nx"}, \code{"auc_R_nx"}, \code{"Q_local"}, \code{"Q_global"}
  \item Spearman correlation of geodesic distances: \code{"correlation"}
  \item Isomorphism of the two networks: \code{"isomorphic"}
  \item GEDEVO Graph Edit Distance: \code{"ged"}
  \item Earth Mover's Distance on orbit counts: \code{"net_emd"}
  \item Genetic Algorithm for aligning small graphs: \code{"robbie_network_score"}
  \item Mantel test p-value: \code{"mantel_pval"}
}}

\item{noisy}{whether or not the metric is noisy or not}
}
\description{
Used for wrapping an evaluation function around a TI method
}
