% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrMBO_objfun_wrapper.R
\name{make_obj_fun}
\alias{make_obj_fun}
\title{Used for wrapping an evaluation function around a TI method}
\usage{
make_obj_fun(method, metrics, extra_metrics, noisy = FALSE)
}
\arguments{
\item{method}{The method to evaluate.}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Coranking of geodesic distances: \code{"mean_R_nx"}, \code{"auc_R_nx"}, \code{"Q_local"}, \code{"Q_global"}
  \item Spearman correlation of geodesic distances: \code{"correlation"}
<<<<<<< HEAD
  \item Isomorphism of the two networks: \code{"isomorphic"}
  \item GEDEVO Graph Edit Distance: \code{"ged"}
  \item Earth Mover's Distance on orbit counts: \code{"net_emd"}
  \item Genetic Algorithm for aligning small graphs: \code{"node_edit_score"}
  \item Edit distance for edges \code{"edge_flip"}
=======
>>>>>>> eacac23edfe5af659b7473e6c36707e52480016e
  \item Mantel test p-value: \code{"mantel_pval"}
  \item Edge flip score: \code{"edge_flip"}
  \item RF MSE: \code{"rf_mse"}, \code{"rf_rsq"}
}}

\item{extra_metrics}{Extra metrics to calculate but not evaluate with.}

\item{noisy}{whether or not the metric is noisy or not}
}
\description{
Used for wrapping an evaluation function around a TI method
}
