% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_metric.R
\name{make_obj_fun}
\alias{make_obj_fun}
\title{Used for wrapping an evaluation function around a TI method}
\usage{
make_obj_fun(method, metrics, noisy = FALSE)
}
\arguments{
\item{method}{the method to wrap an evaluation function around}

\item{metrics}{which metrics to evaluate with;
see \code{\link{calculate_metrics}} for a list of which metrics are available.}

\item{noisy}{whether or not the metric is noisy or not}
}
\description{
Used for wrapping an evaluation function around a TI method
}
