% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{Metadata on metrics}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 7 columns.}
\usage{
metrics
}
\description{
Metadata on metrics
}
\keyword{datasets}
