% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_flip.R
\name{plot_edge_flips}
\alias{plot_edge_flips}
\title{Plotting edge flips}
\usage{
plot_edge_flips(oldadj, newadj)
}
\arguments{
\item{oldadj}{Zouter needs to}

\item{newadj}{fill in these fields}
}
\description{
Plotting edge flips
}
\examples{
net1 <- dyntoy:::generate_toy_milestone_network("linear_long")
net2 <- dyntoy:::generate_toy_milestone_network("trifurcating")
result <- calculate_edge_flip(net1, net2)
plot_edge_flips(result$newadj1, result$oldadj1)
}
