% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_flip.R
\name{plot_edge_flips}
\alias{plot_edge_flips}
\title{Plotting edge flips}
\usage{
plot_edge_flips(oldadj, newadj)
}
\arguments{
\item{oldadj}{Old adjacency matrix}

\item{newadj}{New adjancency matrix}
}
\description{
Plotting edge flips
}
