% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_emdist}
\alias{plot_emdist}
\title{Plot the Earth Mover's distances in a heatmap}
\usage{
plot_emdist(traj, dist, dimred = NULL, ...)
}
\arguments{
\item{traj}{the trajectory (less than 500 cells is recommended)}

\item{dimred}{the dimensionality reduction of the trajectory as produced by \code{\link{plotLearnerData.ti.default}}}

\item{emdist}{the Earth Mover's distances as calculated by \code{\link{emdist}}}
}
\description{
Plot the Earth Mover's distances in a heatmap
}
