% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_celltree.R
\name{run_celltree}
\alias{run_celltree}
\title{Run celltree}
\usage{
run_celltree(counts, method = "maptpx",
  num_topics = num_topics_lower:num_topics_upper, sd_filter = 0.5,
  tot_iter = 1e+06, tolerance = 0.05, width_scale_factor = 1.5)
}
\arguments{
\item{counts}{the counts matrix}

\item{num_topics}{a range of values corresponding to \code{k.topics}}

\item{sd_filter}{a standard deviation filter corresponding to \code{sd.filter}}

\item{tot_iter}{the total number of iterations corresponding to \code{tot.iter}}

\item{tolerance}{the tolerance corresponding to \code{tol}}

\item{width_scale_factor}{the width scale factor corresponding to \code{width.scale.factor}}
}
\description{
This function is a wrapper for \code{\link[cellTree]{compute.lda}} and
\code{\link[cellTree]{compute.backbone.tree}} and finally using
\code{\link{wrap_ti_prediction}} to return the output.
}
