% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_method.R
\name{run_method}
\alias{run_method}
\title{Run method given a task, a method description, and its parameters}
\usage{
run_method(task, method, parameters, suppress_output = TRUE)
}
\arguments{
\item{task}{a task generated by \code{\link{wrap_ti_task_data}}}

\item{method}{a method description}

\item{parameters}{the parameters to apply}

\item{suppress_output}{whether or not to print messages}
}
\description{
Run method given a task, a method description, and its parameters
}
